Do PID control of position with the Sparkfun motor driver.
If our motor driver board changes in the future, we can use motors.h as
a HAL to hide that change from the rest of the code.

Be sure *not* to use pointers when an object really should be owning a
child object. E.g., gunnar should own both motors, while
controlledMotors should just have pointers to them, and controlledMotors
should own both PIDs.

Also, be sure that the correct motors are connected to the correct
output pins on the motor driver board.