Bypass Arduino.mk upload directive.
Running `make upload` might actually compile the code twice, though--
once with the target from Arduino.mk (which will fail to upload), and
again through the IDE. Instead, use `make` or `make all` to check for
errors and warnings, or `make go` to check for errors and upload.

The target `make stop` will compile and upload an empty dummy sketch.