When called on to go() somewhere, loop until arrived.
Wait until the euclidean distance between the set points and the
monitored values is below some threshold, or a timeout has been reached,
whichever comes first. Probably the threshold should be raised (or made
an optional parameter to go()), and the timeout should be lengthened (to
something not likely to be reached except in some yet-unknown failure
modes).