Running-average the heading data.
Also, ms = us/ 1000, not *1000. There are some longs in there that can
probably be converted back to unsigned longs, now.

Tasks in the taskdriver loop should be pointers, so that we can give
them the current time as measured by the taskdriver code. Otherwise, I
think they might just be copies, and so the original tasks are not
changed?