Plot (nearly) arbitrary CSV files.
Other crap can be mixed in with the good lines. Number-of-fields is
taken as either the largest number of elements of any one line when all
lines are split by ',', or as the second command-line argument (after
the path to the .csv file).

Lines with more or fewer fields than this are ignored.

Among unignored lines, the first element must be a time or abscissa (t).
Remaining elements (x, y, z, etc.) are plotted (x vs t, y vs t, etc.)
and according to their place in the line (1, 2, 3, etc.).