Let wiring determine motor sense.
Also, I was having trouble with the second motor not responding to
direction change commands. I traced the problem to a bad connection at
microcontroller digital out pin 23, and applied soldering iron heat
there. This seems to have fixed it, but there's a possibility that the
wire is bad, or the connection is bad at its other end, and it was just
the moving of the wire that fixed it. The connection at pin 23 seemed
mechanically strong. However, applying a multimeter pin there (V test to
ground), and later just a finger, seemed to temporarily fix the problem,
so I think it was in fact a problem right at that solder pad.